+++ apis/appmesh/v1beta1/zz_gatewayroute_types.go
@@ -78,7 +78,7 @@ type ActionRewriteParameters struct {
 type ActionTargetInitParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []TargetVirtualServiceInitParameters `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -87,7 +87,7 @@ type ActionTargetInitParameters struct {
 type ActionTargetObservation struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []TargetVirtualServiceObservation `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -97,7 +97,7 @@ type ActionTargetParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	// +kubebuilder:validation:Optional
@@ -290,7 +290,7 @@ type HTTPRouteActionParameters struct {
 type HTTPRouteActionTargetInitParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []ActionTargetVirtualServiceInitParameters `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -299,7 +299,7 @@ type HTTPRouteActionTargetInitParameters struct {
 type HTTPRouteActionTargetObservation struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []ActionTargetVirtualServiceObservation `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -309,7 +309,7 @@ type HTTPRouteActionTargetParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	// +kubebuilder:validation:Optional
@@ -366,7 +366,7 @@ type HTTPRouteMatchInitParameters struct {
 	Path []HTTPRouteMatchPathInitParameters `json:"path,omitempty" tf:"path,omitempty"`
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	Prefix *string `json:"prefix,omitempty" tf:"prefix,omitempty"`
@@ -387,7 +387,7 @@ type HTTPRouteMatchObservation struct {
 	Path []HTTPRouteMatchPathObservation `json:"path,omitempty" tf:"path,omitempty"`
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	Prefix *string `json:"prefix,omitempty" tf:"prefix,omitempty"`
@@ -412,7 +412,7 @@ type HTTPRouteMatchParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	// +kubebuilder:validation:Optional
@@ -639,7 +639,7 @@ type Http2RouteMatchInitParameters struct {
 	Path []MatchPathInitParameters `json:"path,omitempty" tf:"path,omitempty"`
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	Prefix *string `json:"prefix,omitempty" tf:"prefix,omitempty"`
@@ -660,7 +660,7 @@ type Http2RouteMatchObservation struct {
 	Path []MatchPathObservation `json:"path,omitempty" tf:"path,omitempty"`
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	Prefix *string `json:"prefix,omitempty" tf:"prefix,omitempty"`
@@ -685,7 +685,7 @@ type Http2RouteMatchParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	// +kubebuilder:validation:Optional
@@ -846,7 +846,7 @@ type MatchHostnameParameters struct {
 type MatchInitParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Fully qualified domain name for the service to match from the request.
 	ServiceName *string `json:"serviceName,omitempty" tf:"service_name,omitempty"`
@@ -855,7 +855,7 @@ type MatchInitParameters struct {
 type MatchObservation struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Fully qualified domain name for the service to match from the request.
 	ServiceName *string `json:"serviceName,omitempty" tf:"service_name,omitempty"`
@@ -865,7 +865,7 @@ type MatchParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Fully qualified domain name for the service to match from the request.
 	// +kubebuilder:validation:Optional
@@ -952,30 +952,30 @@ type MatchQueryParameterParameters struct {
 type MatchRangeInitParameters struct {
 
 	// End of the range.
-	End *float64 `json:"end,omitempty" tf:"end,omitempty"`
+	End *int64 `json:"end,omitempty" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
-	Start *float64 `json:"start,omitempty" tf:"start,omitempty"`
+	Start *int64 `json:"start,omitempty" tf:"start,omitempty"`
 }
 
 type MatchRangeObservation struct {
 
 	// End of the range.
-	End *float64 `json:"end,omitempty" tf:"end,omitempty"`
+	End *int64 `json:"end,omitempty" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
-	Start *float64 `json:"start,omitempty" tf:"start,omitempty"`
+	Start *int64 `json:"start,omitempty" tf:"start,omitempty"`
 }
 
 type MatchRangeParameters struct {
 
 	// End of the range.
 	// +kubebuilder:validation:Optional
-	End *float64 `json:"end" tf:"end,omitempty"`
+	End *int64 `json:"end" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
 	// +kubebuilder:validation:Optional
-	Start *float64 `json:"start" tf:"start,omitempty"`
+	Start *int64 `json:"start" tf:"start,omitempty"`
 }
 
 type PathInitParameters struct {
@@ -1077,30 +1077,30 @@ type QueryParameterParameters struct {
 type RangeInitParameters struct {
 
 	// End of the range.
-	End *float64 `json:"end,omitempty" tf:"end,omitempty"`
+	End *int64 `json:"end,omitempty" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
-	Start *float64 `json:"start,omitempty" tf:"start,omitempty"`
+	Start *int64 `json:"start,omitempty" tf:"start,omitempty"`
 }
 
 type RangeObservation struct {
 
 	// End of the range.
-	End *float64 `json:"end,omitempty" tf:"end,omitempty"`
+	End *int64 `json:"end,omitempty" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
-	Start *float64 `json:"start,omitempty" tf:"start,omitempty"`
+	Start *int64 `json:"start,omitempty" tf:"start,omitempty"`
 }
 
 type RangeParameters struct {
 
 	// End of the range.
 	// +kubebuilder:validation:Optional
-	End *float64 `json:"end" tf:"end,omitempty"`
+	End *int64 `json:"end" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
 	// +kubebuilder:validation:Optional
-	Start *float64 `json:"start" tf:"start,omitempty"`
+	Start *int64 `json:"start" tf:"start,omitempty"`
 }
 
 type RewriteHostnameInitParameters struct {
@@ -1221,7 +1221,7 @@ type SpecInitParameters struct {
 	Http2Route []Http2RouteInitParameters `json:"http2Route,omitempty" tf:"http2_route,omitempty"`
 
 	// Priority for the gateway route, between 0 and 1000.
-	Priority *float64 `json:"priority,omitempty" tf:"priority,omitempty"`
+	Priority *int64 `json:"priority,omitempty" tf:"priority,omitempty"`
 }
 
 type SpecObservation struct {
@@ -1236,7 +1236,7 @@ type SpecObservation struct {
 	Http2Route []Http2RouteObservation `json:"http2Route,omitempty" tf:"http2_route,omitempty"`
 
 	// Priority for the gateway route, between 0 and 1000.
-	Priority *float64 `json:"priority,omitempty" tf:"priority,omitempty"`
+	Priority *int64 `json:"priority,omitempty" tf:"priority,omitempty"`
 }
 
 type SpecParameters struct {
@@ -1255,13 +1255,13 @@ type SpecParameters struct {
 
 	// Priority for the gateway route, between 0 and 1000.
 	// +kubebuilder:validation:Optional
-	Priority *float64 `json:"priority,omitempty" tf:"priority,omitempty"`
+	Priority *int64 `json:"priority,omitempty" tf:"priority,omitempty"`
 }
 
 type TargetInitParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []VirtualServiceInitParameters `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -1270,7 +1270,7 @@ type TargetInitParameters struct {
 type TargetObservation struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []VirtualServiceObservation `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -1280,7 +1280,7 @@ type TargetParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	// +kubebuilder:validation:Optional
