+++ apis/apprunner/v1beta1/zz_vpcconnector_types.go
@@ -52,7 +52,7 @@ type VPCConnectorObservation struct {
 	VPCConnectorName *string `json:"vpcConnectorName,omitempty" tf:"vpc_connector_name,omitempty"`
 
 	// The revision of VPC connector. It's unique among all the active connectors ("Status": "ACTIVE") that share the same Name.
-	VPCConnectorRevision *float64 `json:"vpcConnectorRevision,omitempty" tf:"vpc_connector_revision,omitempty"`
+	VPCConnectorRevision *int64 `json:"vpcConnectorRevision,omitempty" tf:"vpc_connector_revision,omitempty"`
 }
 
 type VPCConnectorParameters struct {
